Refactor userland/syscall into _start for new process initialization
